Added puid project identifier argument to the initial deploy script.
Added ssr port to the config so ssr script uses that port to spawn a server.

Updated nginx config on initial deploy.

Removed start argument from git script. Start is the default behaviour.

Added custom nginx config support as it will be included in the main config of the host.

+++ gitflow/index.js
@@ -40,10 +40,10 @@ module.exports = function gitflow(argv) {
 
   // create a temporary file for commit message
   const commitMsgFile = path.join(os.homedir(), 'COMMIT_MESSAGE.txt')
-  if (!argv.message && !fs.existsSync(commitMsgFile)) {
+  if (!argv.message && !fs.existsSync(commitMsgFile) && isFinishing) {
     throw new Error(`Either no -m flag specified and no ${commitMsgFile} file found for commit message.`)
   }
-  if (argv.message) {
+  if (argv.message && isFinishing) {
     const commitMessages = Array.isArray(argv.m) ? argv.m : [argv.m]
     const commitMsgFileContent = commitMessages.reduce(function(memo, msg) {
       memo += msg + "\r\n"
@@ -91,5 +91,7 @@ module.exports = function gitflow(argv) {
     }
   }
 
-  execSync(`rm ${commitMsgFile}`)
+  if (isFinishing) {
+    execSync(`rm ${commitMsgFile}`)
+  }
 }
